requirements.txt
# CoreAgent-2 — Production-Grade Dependencies
# Framework
streamlit>=1.40.0
langgraph>=0.2.0
langgraph-checkpoint-sqlite>=2.0.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
chromadb>=0.5.0

# Document Processing
pypdf>=4.0.0
Pillow>=10.0.0
pytesseract>=0.3.10
pdf2image>=1.17.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.27.0

# Analytics & Visualization
plotly>=5.18.0
pandas>=2.0.0
numpy>=1.24.0

# Local Embeddings (optional — Ollama used by default)
sentence-transformers>=3.0.0
